Clothing Co-Parsing/requirements.txt
gradio
flask
gunicorn
torch
torchvision
numpy
Pillow
opencv-python-headless
requests
tqdm
flask_cors